this fixes prio in MX and in SRV records from being zero on sun-endian platforms
git-svn-id: svn://svn.powerdns.com/pdns/trunk/pdns@773 d19b8d6e-7fed-0310-83ef-9ca221ded41b

+++ pdns/dnswriter.cc
@@ -101,7 +101,7 @@ void DNSPacketWriter::xfr32BitInt(uint32_t val)
 
 void DNSPacketWriter::xfr16BitInt(uint16_t val)
 {
-  int rval=htons(val);
+  uint16_t rval=htons(val);
   uint8_t* ptr=reinterpret_cast<uint8_t*>(&rval);
   d_record.insert(d_record.end(), ptr, ptr+2);
 }
